% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTdesign.R
\name{active}
\alias{active}
\title{active - active bindings for ICTdesign R6 class definitions and ALL children}
\usage{
active()
}
\description{
active - active bindings for ICTdesign R6 class definitions and ALL children
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
