% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catMat}
\alias{catMat}
\title{catMat - function to make a numeric matrix compatible with `cat`}
\usage{
catMat(x = matrix(1:4, 2, 2))
}
\description{
catMat - function to make a numeric matrix compatible with `cat`
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
