% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkCorMat}
\alias{checkCorMat}
\title{checkCorMat}
\usage{
checkCorMat(corMat, cor = TRUE)
}
\description{
checkCorMat
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
