% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r
\name{checkEffectSizesPoly}
\alias{checkEffectSizesPoly}
\title{checkEffectSizes}
\usage{
checkEffectSizesPoly(effectSizes)
}
\description{
checkEffectSizes
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
