% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r
\name{checkPolyICT}
\alias{checkPolyICT}
\title{checkPolyICT}
\usage{
checkPolyICT(effectSizes, corMat, randFxVar)
}
\description{
checkPolyICT
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
