% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT2.R
\name{checkPolyICT2}
\alias{checkPolyICT2}
\title{checkPolyICT2}
\usage{
checkPolyICT2(effectSizes, corMat, randFxVar)
}
\description{
checkPolyICT2
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
