% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.R, R/polyICT2.R
\name{checkRandFxMeanPoly}
\alias{checkRandFxMeanPoly}
\title{checkRandFxMean}
\usage{
checkRandFxMeanPoly(randFxMean)

checkRandFxMeanPoly(randFxMean)
}
\description{
checkRandFxMean

checkRandFxMean
}
\author{
Stephen Tueller \email{stueller@rti.org}

Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
