% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cutY}
\alias{cutY}
\title{cutY - categorize x by a vector of proportions that sum to 1}
\usage{
cutY(y, yCut)
}
\description{
cutY - categorize x by a vector of proportions that sum to 1
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
