% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{ddReport}
\alias{ddReport}
\title{ddReport - function to save descriptives of descriptives to file}
\usage{
ddReport(ddescriptives, file)
}
\arguments{
\item{ddescriptives}{data frame of descriptives of descriptives}
}
\description{
ddReport - function to save descriptives of descriptives to file
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
