% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT2.R
\name{deconstructY}
\alias{deconstructY}
\title{deconstructY - a function to deconstruct Y into
random effects and error}
\usage{
deconstructY(randFx, y, ymean, yvar)
}
\description{
deconstructY - a function to deconstruct Y into
random effects and error
}
