% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{designCheck}
\alias{designCheck}
\title{designCheck - check whether a file is writable and a csv or Rdata file}
\usage{
designCheck(design, file, family, randFxParms, randFxSeed, errorParms,
  errorFUN, errorFamily, errorSeed)
}
\description{
designCheck - check whether a file is writable and a csv or Rdata file
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
