% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designICT.R
\docType{class}
\name{designICT}
\alias{designICT}
\title{\code{designICT} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
designICT
}
\description{
The \code{designICT} class is a parent class to \code{\link{polyICT}} and
other future planned classes such as \code{expICT} for exponential growth
models and \code{logisticICT} for logistic ICT growth models. The
\code{designICT} class contains methods (functions common to all child
classes) and fields (variables common to all child classes) and is not
to be used directly.
}
\section{Methods}{

\describe{

  \item{\code{print}}{

    A method for printing a the primary inputs for an ICT design.

  }

  \item{\code{designCheck}}{

    This method simulates one dataset using the current ICT design and
    plots average trajectories and histograms of the outcome by phase and
    by group. This allows the user to visually check whether the inputs they
    provided match their intendend design. The default is to do this with
    large samples (N=5,000 per group) to show the asypmtotic behavior. For
    comparison, smaller sample sizes can also be used to see the possible
    deviations from the population. If smaller sample sizes are used, this
    should be repeated several times to prevent over-interpretation of one
    sample. See the examples in \code{\link{polyICT}}.

    \code{file} Character. A file name for saving the resulting data. The
    default is NULL, in which case the data are not saved.

    \code{ylim} Numeric vector of length 2. Limits for the y-axis. This is
    useful when variance is high and/or effect sizes are small, making it
    difficult to visualize average change over time. The default is NULL, in
    which case default limits are used.

    \code{fitMod} Logical. Should the model implied by the design be fit to
    the simulated data? The default is FALSE.

    - \code{seed} Numeric. A random seed for enabling replication. The default
    is 123.

    \code{npg} Numeric (integer). The number per group. The default is
    n=5,000 per group. If a small value is used, the user should repeat the
    check with several different seeds. See the examples in
    \code{\link{polyICT}}.
  }

}
}

\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
