% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designICT.R
\name{.active}
\alias{.active}
\title{active - active bindings for designICT R6 class definitions and it heirs}
\usage{
.active()
}
\description{
active - active bindings for designICT R6 class definitions and it heirs
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
