% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{.arima.sim}
\alias{.arima.sim}
\title{.arima.sim - function to run arima.sim with potentially extraneous arguments
intended for a gamlss.family distribution passed to rand.gen}
\usage{
.arima.sim(seed, doPlot = TRUE, doStats = TRUE, ...)
}
\description{
.arima.sim - function to run arima.sim with potentially extraneous arguments
intended for a gamlss.family distribution passed to rand.gen
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
