% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.checkFam}
\alias{.checkFam}
\title{.checkFam - check whether \code{fam} is a gamlss.family object and
whether its parameters in \code{parms} are legitimate}
\usage{
.checkFam(fam, parms)
}
\description{
.checkFam - check whether \code{fam} is a gamlss.family object and
whether its parameters in \code{parms} are legitimate
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
