% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.checkFile}
\alias{.checkFile}
\title{.checkFile - check whether a file is writable and a csv or Rdata file}
\usage{
.checkFile(file, prompt = TRUE)
}
\description{
.checkFile - check whether a file is writable and a csv or Rdata file
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
