% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.doIt}
\alias{.doIt}
\title{.doIt - helper function for .doLapply for ignoring extra arguments, written
to allow calls to \code{\link{gamlss.family}} distributions but ignore unused
parameters}
\usage{
.doIt(X, .fcn = "qNO", ...)
}
\arguments{
\item{X}{a vector of probabilities to be passed to a quantile function
like \code{\link{qNO}} in \code{\link{gamlss}} or \code{\link{qnorm}}.}

\item{.fcn}{see \code{\link{doCall}} in the `R.utils` package.}
}
\description{
.doIt - helper function for .doLapply for ignoring extra arguments, written
to allow calls to \code{\link{gamlss.family}} distributions but ignore unused
parameters
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
