% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.err}
\alias{.err}
\title{.err - function to generate errors in .active}
\usage{
.err(x, self)
}
\description{
.err - function to generate errors in .active
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
