% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.hl}
\alias{.hl}
\title{.hl - horizontal line for printing to the console}
\usage{
.hl()
}
\description{
.hl - horizontal line for printing to the console
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
