% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.makeSeeds}
\alias{.makeSeeds}
\title{.makeSeeds - make a vector of random seeds}
\usage{
.makeSeeds(seed, S)
}
\description{
.makeSeeds - make a vector of random seeds
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
