% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMakers.R
\name{.polyData}
\alias{.polyData}
\title{.polyData - function to simulate polynomial growth data, used by the polyICT
class}
\usage{
.polyData(
  seed = 123,
  n,
  nObs,
  mu,
  Sigma,
  self,
  dM,
  dM01,
  rFxVr,
  propErrVar,
  group = NULL
)
}
\description{
.polyData - function to simulate polynomial growth data, used by the polyICT
class
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
