% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.randFxCorMatPop}
\alias{.randFxCorMatPop}
\title{.randFxCorMatPop - correlation matrix populator, see the `designCheck` method in \code{\link{polyICT}}}
\usage{
.randFxCorMatPop(phaseNames, groupNames, randFxCorMat, wn = c("p", "g", "n"))
}
\description{
.randFxCorMatPop - correlation matrix populator, see the `designCheck` method in \code{\link{polyICT}}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
