% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.statNameParse}
\alias{.statNameParse}
\title{.statNameParse - used by samplingDist()}
\usage{
.statNameParse(statName)
}
\description{
.statNameParse - used by samplingDist()
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
