% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.totalVar}
\alias{.totalVar}
\title{.totalVar - get the total variance for simulated data}
\usage{
.totalVar(randFxCovMat, propErrVar, designMat, randFxMean, nObservations, n)
}
\description{
.totalVar - get the total variance for simulated data
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
