% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.update}
\alias{.update}
\title{.update - function to update multiple objects when any
one object is updated}
\usage{
.update(..., anICT)
}
\description{
.update - function to update multiple objects when any
one object is updated
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
