% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{errActive}
\alias{errActive}
\title{errActive Active bindings for Err and its heirs}
\usage{
errActive()
}
\description{
errActive Active bindings for Err and its heirs
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
