% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r
\name{expectedVar}
\alias{expectedVar}
\title{expectedVar}
\usage{
expectedVar(covMat, designMat, variances, effectSizes, nObservations, n)
}
\description{
expectedVar
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
