% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdesignICT}
\alias{getdesignICT}
\title{getdesignICT - create the fixed effects design matrix for n=1}
\usage{
getdesignICT(phases = makePhase(), phaseNames = NULL, maxRandFx = 2,
  design = "polyICT")
}
\description{
getdesignICT - create the fixed effects design matrix for n=1
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
