% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getICTdesign}
\alias{getICTdesign}
\title{getICTdesign - create the fixed effects design matrix for n=1}
\usage{
getICTdesign(phases = makePhase(), polyOrder = 2, design = "polyICT")
}
\description{
getICTdesign - create the fixed effects design matrix for n=1
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
