% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIinputMat.R
\name{gui1}
\alias{gui1}
\title{Translator for GUI output version April 2020}
\usage{
gui1(inputMat, propErrVar)
}
\arguments{
\item{inputMat}{An inputmatrix from a \code{\link{polyICT}} object.}
}
\description{
Translator for GUI output version April 2020
}
