% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIinputMat.R
\name{GUIinputMat}
\alias{GUIinputMat}
\title{Function to translate input matrices from the PersonAlytics GUI}
\usage{
GUIinputMat(inputMat, translator = gui1)
}
\arguments{
\item{inputMat}{An input matrix created by the PersonAlytics GUI, which
may or may not have been edited by a user.}

\item{translator}{The translator version. Currently only \code{gui1} is
available, more will be added if needed.}
}
\description{
Function to translate input matrices from the PersonAlytics GUI
}
\examples{

data(GUIinputMatExample)
inputMat <- GUIinputMat(GUIinputMatExample)

myPolyICT <- polyICT$new(
  groups            = c(group1=10, group2=20)          ,
  phases            = makePhase()                      ,
  propErrVar        = c(randFx=.5, res=.25, mserr=.25) ,
  randFxOrder       = 1                                ,
  randFxCor         = 0.2                              ,
  randFxVar         = c(1, 1)                          ,
  error             = armaErr$new()                    ,
  merror            = armaErr$new(list())              ,
  ySD               = 15                               ,
  yMean             = 100                              ,
  )

myPolyICT$inputMat <- inputMat

\dontrun{
myPolyICT$designCheck()
}


}
