% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designICT.R
\name{ICTactive}
\alias{ICTactive}
\title{active - active bindings for designICT R6 class definitions and it heirs}
\usage{
ICTactive()
}
\description{
active - active bindings for designICT R6 class definitions and it heirs
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
