% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTdesign.r
\docType{data}
\name{ICTdesign}
\alias{ICTdesign}
\title{ICTdesign}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ICTdesign
}
\description{
ICTdesign
}
\examples{
# create a basic ABA design
basicICT <- ICTdesign$new(n=10, phases=makePhase(), propErrVar=.5)
basicICT
basicICT$n <- 20
basicICT$n <- 'A'
basicICT$propErrVar <- 2
}
\keyword{datasets}
