% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simICTdata.R
\name{ICTerror}
\alias{ICTerror}
\title{ICTerror}
\usage{
ICTerror(design = polyICT$new(), errorParms = list(ar = c(0.5), ma =
  c(0)), errorFUN = arima.sim, seed = 1, ...)
}
\description{
ICTerror
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
