% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpowerSim.R
\name{ICTpowerSim}
\alias{ICTpowerSim}
\title{ICTpowerSim - automate a series of parametric bootstrap power simulations.}
\usage{
ICTpowerSim(
  designs,
  pReportName = "ICTpowerSimResults",
  B = 1000,
  seed = 1,
  save = NULL,
  alpha = 0.05,
  cores = parallel::detectCores() - 1,
  standardize = list(dv = TRUE, ivs = FALSE, byids = TRUE),
  dotar = FALSE,
  ...
)
}
\arguments{
\item{designs}{A named list of designs.
in \code{\link{ICTpower}}. The names in the named list will be passed to
the parameter \code{file} in \code{\link{ICTpower}}.}

\item{pReportName}{Character. A name used to create output files. The default
is \code{"ICTpowerSimResults"}}

\item{B}{Numeric (integer). The number of replicated data sets per \code{design}
in \code{designs}.}

\item{save}{\code{NULL} or character. Should the data be saved? If NULL
(the default) raw data will not be saved. Other
options include \code{save='csv'} and \code{save='RData'}. This will be
combined with the names of \code{designs} to pass to the \code{file} option
in \code{\link{ICTpower}}.}

\item{alpha}{See \code{\link{ICTpower}}.}

\item{cores}{See \code{\link{ICTpower}}.}

\item{standardize}{See \code{\link{ICTpower}}.}

\item{dotar}{Logical. The default is \code{FALSE}. Should the data files
(if requested, see \code{save}) and analysis results be saved in a *.tar
file with the same name as \code{pReportName} and then delete the unzipped
copies of these files? This is suggested wheth the number of designs is
large. The summary of power across all conditions will no be included in the
*.tar archive. It is strongly encourage that a test run with only three
designs and B=3 is attempted at first to make sure archiving is done
correctly. If the path names resulting from your working directory are
too long, archiving will fail with the error "storing paths of more than 100
bytes is not portable". See \code{\link{tar}}.}

\item{Seed}{Numeric (integer). The random seed for ensuring the simulation
can be replicated.}
}
\description{
ICTpowerSim - automate a series of parametric bootstrap power simulations.
}
\examples{

\dontrun{

# set up two designs
designs <- list()
designs[["defaultsN20"]] <- polyICT$new()
designs[["defaultsN40"]] <- designs[["defaultsN20"]]$clone(deep=TRUE)
designs[["defaultsN40"]]$update(groups = c(group1=40, group2=40))

# run the simulation (without saving output, console print only)
ICTpowerSim(designs, pReportName  = "ex1", B=3)

# rerun with a finite population correction
ICTpowerSim(designs, pReportName  = "ex1fpc", B=3, seed=2, fpc=100)

}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
