% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ICTsetup}
\alias{ICTsetup}
\alias{makePhase}
\alias{polyICTsetup}
\alias{ICTviz}
\title{functions for setting up ICT designs.}
\usage{
makePhase(nObsPerPhase = c(10, 20, 10), phaseNames = c("phase1",
  "phase2", "phase3"))

polyICTsetup(phases = makePhase(), nGroups = 1, effectSizes = NULL,
  DIST = "NO", parms = list(mu = 0, sigma = 1), BLrange = c(-3, -2))

ICTviz(DIST = "NO", parms = list(mu = 0, sigma = 1, nu = 2, tau = 2),
  designMatrix = NULL, BLrange = NULL)
}
\arguments{
\item{nObsPerPhase}{Numeric vector. The length is the number of phases
and each entry is the number of observations per phase.}

\item{phaseNames}{Numeric or character vector. The names of the phases,
e.g., phase 0 and phase 1, or phases "A", "B", and "C". Phase names must
be unique, i.e., even if your design is ABA, you should use 1, 2, 3, or
"A", "B", "C".}

\item{phases}{List. The length of phases is the number of phases. Each
item in the list is a vector repeating the phase name or number as many
times as there are time points within that phase. See \code{\link{makePhase}}.}

\item{nGroups}{Integer. The number of groups.}

\item{effectSizes}{See \code{polyICT2}. If \code{effectSizes} is
provided, \code{nGroups} will be set to \code{length(effectSizes)}.}

\item{DIST}{Character. A \code{\link{gamlss.family}} distribution.}

\item{parms}{Named list. The parameters for \code{DIST}. \code{parms} should
be from length 1 to 4 with possible names including 'mu', 'sigma', 'nu', and
'tau'.}

\item{BLrange}{Numeric vector. The range of possible values at baseline. This
range should be smaller than, and within, the theoretical range of the
distribution specified by \code{DIST}.}

\item{designMatrix}{Matrix. See \code{\link{polyICTsetup}}.}
}
\value{
For \code{makePhase}, a list of phase names replicated by the number
of observations per phase.

For \code{polyICTsetup},
a data.frame with a `time` and `phase` variable, plus the following pairs of
variables each group:
  \itemize{
    \item lower_Group#. The lower bound at each time point.
    \item upper_Group#. The upper bound at each time point.
  }
where the `#` is replaced by the group number. This naming convention is required
for other functions in the package.

Lower/upper bounds give the truncated range of the parent distribution
from which observed values are sampled. ICTs often look at subpopulations whose
scores lie on one end of the distribution, and a study intervetion aims to move
them toward scores that are more normal.

For example, in a three phase study with 5 time points in each phase. The first
phase is a static baseline, the second phase occurs after an intervention which
raises scores, and the third phase shows linear decay back to the original
bounds after the removal of the intervention:

\tabular{rrrr}{
time \tab phase \tab lower_Group1 \tab upper_Group1 \cr
   1 \tab     A \tab           -2 \tab            0 \cr
   2 \tab     A \tab           -2 \tab            0 \cr
   3 \tab     A \tab           -2 \tab            0 \cr
   4 \tab     A \tab           -2 \tab            0 \cr
   5 \tab     A \tab           -2 \tab            0 \cr
   6 \tab     B \tab           -1 \tab            1 \cr
   7 \tab     B \tab           -1 \tab            1 \cr
   8 \tab     B \tab           -1 \tab            1 \cr
   9 \tab     B \tab           -1 \tab            1 \cr
  10 \tab     B \tab           -1 \tab            1 \cr
  11 \tab     A \tab           -1 \tab            1 \cr
  12 \tab     A \tab        -1.25 \tab         0.75 \cr
  13 \tab     A \tab        -1.50 \tab         0.50 \cr
  14 \tab     A \tab        -1.75 \tab         0.25 \cr
  15 \tab     A \tab           -2 \tab            0
}

In this example, the modal value will be halfway between the lower and upper
bound. The mean and median will be determined by the distribution selected
by the user for data sampling and the truncated range of that distribution
defined by the lower and upper bounds.

This matrix can be constructed by the user, the polyICTsetup function simply
creates a shell for the user to fill in.

For \code{ICTviz}, if \code{DIST} and \code{parms} are given
(and optionally,
\code{BLrange}), a density plot for the \code{DIST} is plotted. If
\code{BLrange} is given, lines indicating the truncated part of the distribution
from which baseline values will be sampled are shown.

Additionally,
if \code{designMatrix} is provided, a visualization of the study desgin will
be plotted next to the density plot.
}
\description{
functions for setting up ICT designs.
}
\note{
\code{makePhase} is used to generate a `phase` variable.

\code{polyICTsetup} is used to create a polynomial ICT design.

\code{ICTviz} visualizes ICT sampling and resulting designs.
}
\examples{
makePhase()

# view an example study design matrix that could have been initiated
# with `polyICTsetup`
g2exampleDesignMatrix

# visualize this design using truncated standard normal
ICTviz(DIST = 'NO', parms=list(mu=0, sigma=1),
       designMatrix = g2exampleDesignMatrix)

\dontrun{
# two phase example with 5 time points each
designMatrix <- polyICTsetup(makePhase(c(5,5), c("A","B")), nGroup=1)

# view the empty matrix
designMatrix

# visualize an observed data distribution density plot
ICTviz()

# now the user needs to manually fill in the design matrix, values for
# the lower/upper bounds are taken from the x-axis of the
designMatrix <- edit(designMatrix)
}
# standard normal distribution with baseline values sampled from -3 to -2
ICTviz(DIST = 'NO', parms=list(mu=0, sigma=1), BLrange = c(-3, -2))

ICTviz(DIST = 'NO', parms=list(mu=0, sigma=1),
       designMatrix = g2exampleDesignMatrix, BLrange = c(-2, -0))
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
