% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpowerSim.R
\name{ICTsimSetup}
\alias{ICTsimSetup}
\title{ICTsimSetup - function to automate setting up \code{designs} to be passed to
\code{\link{ICTpowerSim}}, including splitting up the designs over sevar folders
so the simulation can be run in chunks or on multiple computers.}
\usage{
ICTsimSetup(seed, nL, phasesL, propErrVarL, effectSizesL, corMatL,
  randFxVarL, B = 1000, nFolders = NULL, ICTpowerSimOptions = "")
}
\arguments{
\item{seed}{Numeric. A random seed for this simulation.}

\item{nL}{Numeric list. A list of sample sizes. See \code{n} in
\code{\link{ployICT}}.}

\item{phasesL}{A list of phases. See \code{n} in
\code{\link{ployICT}}.}

\item{propErrVarL}{Numeric List. See \code{n} in
\code{\link{ployICT}}.}

\item{effectSizesL}{A list of effect sizes. See \code{n} in
\code{\link{ployICT}}.

 @param corMatL A list of numeric correlation matrices. See \code{n} in
\code{\link{ployICT}}.}

\item{randFxVarL}{A list of numeric vectors. See \code{n} in
\code{\link{ployICT}}.}

\item{nFolders}{Numeric. The number of folders to divide the simulation into.
The default is \code{NULL}, in which case the user is prompted for input at
the console after the total number of conditions is computed. To prevent
folders that are too large, it is reccomended that \code{nFolders} not
exceed 100, especially on Windows systems.}

\item{ICTpowerSimOptions}{Character. Other options to be passed to the
function \code{\link{ICTpowerSim}}. For example,}

\item{muFUN}{}
}
\description{
ICTsimSetup - function to automate setting up \code{designs} to be passed to
\code{\link{ICTpowerSim}}, including splitting up the designs over sevar folders
so the simulation can be run in chunks or on multiple computers.
}
\examples{
\dontrun{
seed <- 1234
nL <- list(10, 20)
phasesL <- list( makePhase(c(10,10), c(0,1)), makePhase(c(5,15), c(0,1)) )
propErrVarL <- list(.5, .75)
effectSizesL <- list( list(randFx=list(intercept=0, slope=.25),
                      fixdFx=list(phase=.5, phaseTime=.25)),
                      list(randFx=list(intercept=0, slope=.5),
                      fixdFx=list(phase=.5, phaseTime=.25)))
corMatL <- list(matrix(c(1,.2,.2,1), 2, 2), matrix(c(1,.6,.6,1), 2, 2))
randFxVarL <- list(c(1, .1), c(1, .2))
B <- 3
nFolders <- 4
ICTpowerSimOptions <-"cores = 4"
ICTsimSetup(seed    ,
  nL                ,
  phasesL           ,
  propErrVarL       ,
  effectSizesL      ,
  corMatL           ,
  randFxVarL        ,
  B                 ,
  nFolders          ,
  ICTpowerSimOptions)
}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
