% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ICTviz}
\alias{ICTviz}
\title{ICTviz - a function to create the phase variable
and select mean values}
\usage{
ICTviz(DIST = "NO", parms = list(mu = 0, sigma = 1, nu = 2, tau = 2),
  designMatrix = NULL)
}
\description{
ICTviz - a function to create the phase variable
and select mean values
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
