% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{makeBounds}
\alias{makeBounds}
\title{makeBounds - function to get the bounds from the baseline bounds and
vectors of effect sizes}
\usage{
makeBounds(designMatrix, fam, parms, BLrange)
}
\description{
Note that the bounds are approximate on the assumptiong that fixed width
bounds have sufficiently similar SD
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
