% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMaker.R
\name{makeDesign}
\alias{makeDesign}
\title{makeDesign -}
\usage{
makeDesign(
  randFxOrder,
  phases,
  groups,
  propErrVar,
  randFxVar,
  randFxCor,
  design = "polyICT",
  makeInputMat = TRUE,
  self = NULL,
  isNew = NULL
)
}
\description{
makeDesign -
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
