% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designMaker.R
\name{makeDesignMat}
\alias{makeDesignMat}
\title{makeDesignMat - create the fixed effects design matrix for n=1}
\usage{
makeDesignMat(
  phases = makePhase(),
  phaseNames = NULL,
  maxRandFx = 2,
  design = "polyICT"
)
}
\description{
makeDesignMat - create the fixed effects design matrix for n=1
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
