% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMakers.R
\name{makeEta}
\alias{makeEta}
\title{makeEta - a wrapper for mvrnorm that currenty does nothing else but set the
seed and call mvrnorm. Retained as we may need to extend its functionality}
\usage{
makeEta(n, mu, Sigma, seed = 1)
}
\arguments{
\item{n}{See \code{\link{mvrnorm}}.}

\item{mu}{See \code{\link{mvrnorm}}.}

\item{Sigma}{See \code{\link{mvrnorm}}.}

\item{seed}{Default 1. A random seed for replication.}
}
\description{
makeEta - a wrapper for mvrnorm that currenty does nothing else but set the
seed and call mvrnorm. Retained as we may need to extend its functionality
}
\examples{
Y  <- makeEta(1000, c(0,0), matrix(c(1,.3,.3,1), 2, 2))
pairs(Y)
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
