% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{makeMeans}
\alias{makeMeans}
\title{makeMeans - a function to create mean values given within-phase
growth model parameters (i, s, q, etc.; see \code{effectSizes} in
\code{polyICT2}).}
\usage{
makeMeans(phaseParms = c(100, 0.5, -0.25), nObs = 10)
}
\description{
makeMeans - a function to create mean values given within-phase
growth model parameters (i, s, q, etc.; see \code{effectSizes} in
\code{polyICT2}).
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
