% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSeeds}
\alias{makeSeeds}
\title{seeds - make a vector of random seeds}
\usage{
makeSeeds(seed, S)
}
\description{
seeds - make a vector of random seeds
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
