% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{mse}
\alias{mse}
\title{mse - mean squared error & mean absolute error}
\usage{
mse(x)
}
\arguments{
\item{x}{data.frame with columns \code{yhat} (the estimates) and \code{y}
(the population value which should generally have zero variance).}
}
\description{
mse - mean squared error & mean absolute error
}
