% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{mseb}
\alias{mseb}
\title{mseb - version for bootstraping}
\usage{
mseb(x, i)
}
\arguments{
\item{x}{See \code{\link{mse}}}

\item{i}{The index for resampling}
}
\description{
mseb - version for bootstraping
}
\keyword{internal}
