% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r
\docType{class}
\name{polyICT}
\alias{polyICT}
\title{\code{polyICT} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
polyICT
}
\value{
\item{n               }{The number of participants (see `Fields`).}
\item{phases          }{The phases of the study (see `Fields`).}
\item{propErrVar      }{The proportion of error variance (see `Fields`).}
\item{effectSizes     }{The fixed effects effect sizes (see `Fields`).}
\item{corMat          }{The correlation matrix of the random effects (see `Fields`).}
\item{randFxVar       }{The variance of the random effects (see `Fields`).}
\item{muFUN           }{A function for transforming random effects means (see `Fields`).}
\item{SigmaFun        }{A function for constructing the covariance matrix (see `Fields`).}
\item{polyOrder       }{The order of the study. For example, a linear model
would be of order 1, and a quadratic model would be order 2.}
\item{designMat       }{The design matrix for one participant showing the structure
of study timing and phases.}
\item{covMat          }{The covariance matrix constructed using \code{corMat},
\code{randFxVar}, and \code{SigmaFun}.}
\item{nObservations   }{The number of observations per participant.}
\item{variances       }{Partition of the total variance into that due to
random effects and that due to error variance at time = 1.}
\item{expectedVariance}{The expected variances across all time points. This will
not match the variance of the simulated data unless n is large. See the \code{checkDesign}
parameter in \code{\link{ICTpower}}.}
\item{unStdEffects    }{The unstandardized effect sizes constructed from the total
\code{expectedVariance} and the \code{effectSizes}.}
}
\description{
\code{polyICT} class generator
}
\section{Fields}{

\describe{
\item{\code{n}}{Numeric (integer). The number of participants. Default is 10.}

\item{\code{phases}}{List. Each phase in one item in the list with the phase name
repeated for the number of time points in the phase. For example, an "ABA"
study with 5 time points each would be \code{list(rep("A", 5), rep("B", 5),
rep("A", 5))}. See also the function \code{\link{makePhase}}. Default is
\code{makePhase()}.}

\item{\code{propErrVar}}{Numeric. The propotion of total variance that is error
variance. Default is .75.}

\item{\code{effectSizes}}{A length 2 list with the names `randFx` and `fixdFx`. Each
are themselves a list with the coefficients for a polynomial growth model.
The length of `randFx` and `fixdFx` corresponds to the order of the model,
where an order of 2 would be a linear model and an order of 3 would be a
quadratic model. An example of a quadratic model would be
list(randFx = list(intercept = 0, slope = .5, quad = .2),
     fixdFx = list(phase = 1, phaseSlope = .2, phaseQuad = .1))}

\item{\code{corMat}}{Numeric matrix. A symmetric correlation matrix with a dimension
equal to the order of the model. For example, a quadratic model would
correpspond to a 3x3 matrix. The diagonal elements must equal 1, the off
diagonal elements must be between -1 and +1, and the matrix must be
invertable.}

\item{\code{randFxVar}}{Numeric vector. A vector of the same length as the order of
the polynomial
model containing the variances of the random effects. For example, in a
quadratic model, `randFxVar` would be length three, the first element would
be the intercept variance, the second element would be the slope variance,
and the third element would be the variance of the quadratic term.}

\item{\code{SigmaFun}}{An R function. A function to convert \code{corMat} and
\code{randFxVar} into the covariance matrix \code{covMat}. Default is
\code{\link{cor2cov}}.}
}}

\examples{

# produce a simple ICT design
defaultPolyICT <- polyICT$new()

# print a summary
defaultPolyICT

# view the fields that are generated by `$new()` but cannot be changed by
# the user
defaultPolyICT$polyOrder
defaultPolyICT$designMat
defaultPolyICT$covMat
defaultPolyICT$nObservations
defaultPolyICT$variances
defaultPolyICT$expectedVariances
defaultPolyICT$unStdEffects
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
