% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{popSampComp}
\alias{popSampComp}
\title{popSampComp - compare non-parametric bootstrap estimates to population}
\usage{
popSampComp(paout, file)
}
\arguments{
\item{paout}{data.frame produced by \code{\link{PersonAlytic}}.}

\item{file}{character. A file name with no extentsion.}
}
\description{
popSampComp - compare non-parametric bootstrap estimates to population
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
