% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{powerReport}
\alias{powerReport}
\title{powerReport - print power results to screen and to a file}
\usage{
powerReport(paout, alpha, file)
}
\description{
powerReport - print power results to screen and to a file
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
