% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randFxCorMatPop}
\alias{randFxCorMatPop}
\title{randFxCorMatPop - correlation matrix populator, see \code{\link{checkPolyICT}}}
\usage{
randFxCorMatPop(phaseNames, groupNames, randFxCorMat, wn = c("p", "g",
  "n"))
}
\description{
randFxCorMatPop - correlation matrix populator, see \code{\link{checkPolyICT}}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
