% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.r
\name{readinteger}
\alias{readinteger}
\title{readinteger - helper function for studySetup}
\usage{
readinteger(prompt, nth = "")
}
\description{
readinteger - helper function for studySetup
}
\author{
adapted from http://www.rexamples.com/4/Reading%20user%20input
}
\keyword{internal}
