% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{samplingDist}
\alias{samplingDist}
\title{samplingDist - function to plot the sampling distributions of
\code{\link{PersonAlytic}} output produce by \code{\link{ICTpower}}.}
\usage{
samplingDist(file, design)
}
\description{
samplingDist - function to plot the sampling distributions of
\code{\link{PersonAlytic}} output produce by \code{\link{ICTpower}}.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
