% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{sigmaMaker}
\alias{sigmaMaker}
\title{sigmaMaker - create an AR(1) error covariance matrix}
\usage{
sigmaMaker(n = 20, phi = 0.5, sigma2 = 0.75)
}
\description{
sigmaMaker - create an AR(1) error covariance matrix
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
