% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simICTdata.R
\name{simICTdata}
\alias{simICTdata}
\title{simICTdata - this is a user interface function for one dataset,
switching between different designs so the function can be updated}
\usage{
simICTdata(design = polyICT$new(), randFxParms = list(mu = 0.5, sigma =
  1), family = "qNO", randFxSeed = 11, errorParms = list(ar = c(0.5),
  ma = c(0)), errorFUN = arima.sim, errorSeed = 21, B = 100,
  cores = parallel::detectCores() - 1, file = NULL,
  designCheck = FALSE)
}
\description{
simICTdata - this is a user interface function for one dataset,
switching between different designs so the function can be updated
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
