% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpowerSim.R
\name{simStatus}
\alias{simStatus}
\title{simStatus}
\usage{
simStatus(studyName, studyDirectory = getwd())
}
\arguments{
\item{studyName}{}

\item{studyDirectory}{Character. The full path to the directory the
simulation study is being run in}
}
\description{
simStatus
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
