% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{studySetup}
\alias{studySetup}
\title{studySetup - function to help users set up a study design matrix}
\usage{
studySetup(phases = makePhase(), nGroups = 1)
}
\arguments{
\item{phases}{List. The length of phases is the number of phases. Each
item in the list is a vector repeating the phase name or number as many
times as there are time points within that phase. See \code{\link{makePhase}}.}

\item{nGroups}{Integer. The number of groups.}
}
\value{
A data.frame with a `time` and `phase` variable, plus the following pairs of
variables each group:
  \itemize{
    \item lower_Group#. The lower bound at each time point.
    \item upper_Group#. The upper bound at each time point.
  }
where the `#` is replaced by the group number. This naming convention is required
for other functions in the package.

Lower/upper bounds give the truncated range of the parent distribution
from which observed values are sampled. ICTs often look at subpopulations whose
scores lie on one end of the distribution, and a study intervetion aims to move
them toward scores that are more normal.

For example, in a three phase study with 5 time points in each phase. The first
phase is a static baseline, the second phase occurs after an intervention which
raises scores, and the third phase shows linear decay back to the original
bounds after the removal of the intervention:

\tabular{rrrr}{
time \tab phase \tab lower_Group1 \tab upper_Group1 \cr
   1 \tab     A \tab           -2 \tab            0 \cr
   2 \tab     A \tab           -2 \tab            0 \cr
   3 \tab     A \tab           -2 \tab            0 \cr
   4 \tab     A \tab           -2 \tab            0 \cr
   5 \tab     A \tab           -2 \tab            0 \cr
   6 \tab     B \tab           -1 \tab            1 \cr
   7 \tab     B \tab           -1 \tab            1 \cr
   8 \tab     B \tab           -1 \tab            1 \cr
   9 \tab     B \tab           -1 \tab            1 \cr
  10 \tab     B \tab           -1 \tab            1 \cr
  11 \tab     A \tab           -1 \tab            1 \cr
  12 \tab     A \tab        -1.25 \tab         0.75 \cr
  13 \tab     A \tab        -1.50 \tab         0.50 \cr
  14 \tab     A \tab        -1.75 \tab         0.25 \cr
  15 \tab     A \tab           -2 \tab            0
}

In this example, the modal value will be halfway between the lower and upper
bound. The mean and median will be determined by the distribution selected
by the user for data sampling and the truncated range of that distribution
defined by the lower and upper bounds.

This matrix can be constructed by the user, the studySetup function simply
creates a shell for the user to fill in.
}
\description{
studySetup - function to help users set up a study design matrix
}
\examples{

# view an example study design matrix that could have been initiated
# with `studySetup`
g2exampleDesignMatrix

# visualize this design using truncated standard normal
ICTviz(DIST = 'NO', parms=list(mu=0, sigma=1),
       designMatrix = g2exampleDesignMatrix)

\dontrun{
# two phase example with 5 time points each
designMatrix <- studySetup(makePhase(c(5,5), c("A","B")), nGroup=1)

# view the empty matrix
designMatrix

# visualize an observed data distribution density plot
ICTviz()

# now the user needs to manually fill in the design matrix, values for
# the lower/upper bounds are taken from the x-axis of the
designMatrix <- edit(designMatrix)
}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
