% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\docType{class}
\name{armaErr}
\alias{armaErr}
\title{\code{armaErr} class generator using \code{\link{arima.sim}}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
armaErr
}
\description{
\code{armaErr} class generator using \code{\link{arima.sim}}
}
\section{Fields}{

\describe{
\item{\code{parms}}{The parameters for \code{FUN}.}

\item{\code{errVar}}{The error variance.}

\item{\code{FUN}}{The function for simulating error terms.}

\item{\code{fam}}{A \code{\link{gamlss.family}} family for the error term distribution.
After data are simulated using \code{FUN} and its \code{parms}, it can be
transformed to have a distribution different from those available to \code{FUN}.}

\item{\code{famParms}}{Parameters to be passed to \code{fam}, see \code{\link{gamlss.family}}.}
}}

\examples{

# show that r* functions from gamlss.family distribution functions can be
# passed to arima.sim
errBeta <- arima.sim(list(ar=c(.5)), 1000, innov = rBE(1000, mu=.5, sigma=.2))

# note that the rand.gen option only allows teh defaults for the function
# passed to rand.gen
#errBeta <- arima.sim(list(ar=c(.5)), 1000, rand.gen = rBE)
# that said, rand.gen's parameters can be passed via ...
#errBeta <- arima.sim(list(ar=c(.5)), 1000, rand.gen = rBE, mu=.9, sigma=.2)

# note that even though beta innovations are used, the resulting data is not
# constrained to be in (0,1)
hist(errBeta)
plot(errBeta)
auto.arima(ts(errBeta))
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
