% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.R, R/polyICT2.R
\name{checkEffectSizesPoly}
\alias{checkEffectSizesPoly}
\title{checkEffectSizes}
\usage{
checkEffectSizesPoly(effectSizes)

checkEffectSizesPoly(effectSizes)
}
\description{
checkEffectSizes

checkEffectSizes
}
\author{
Stephen Tueller \email{stueller@rti.org}

Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
