% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.R
\name{checkPolyICT}
\alias{checkPolyICT}
\title{checkPolyICT}
\usage{
checkPolyICT(randFxMean, phases, randFxCorMat, randFxVar)
}
\description{
checkPolyICT
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
