% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designICT.R
\docType{class}
\name{designICT}
\alias{designICT}
\title{\code{designICT} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
designICT
}
\description{
\code{designICT} class generator
}
\examples{
# create a basic ABA design
basicICT <- designICT$new(n=10, phases=makePhase(), propErrVar=.5)
basicICT
basicICT$n <- 20
basicICT$n <- 'A'
basicICT$propErrVar <- 2
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
