% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{doLapply}
\alias{doLapply}
\title{doLapply - lapply with \code{\link{doCall}} to ignore extra arguments to
\code{\link{gamlss.family}} distributions}
\usage{
doLapply(Yp, .fcn = "qNO", ...)
}
\arguments{
\item{Yp}{X a vector of probabilities to be based to a quantile function}

\item{.fcn}{A quantile function like \code{\link{qN0}} in \code{\link{gamlss}}
or \code{\link{qnorm}}. See also \code{\link{doCall}} in the `R.utils` package}

\item{...}{other options passed to \code{\link{gamlss.family}} distribution
functions, most commonly mu, sigma, nu, and tau.}
}
\description{
doLapply - lapply with \code{\link{doCall}} to ignore extra arguments to
\code{\link{gamlss.family}} distributions
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
