% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.randFxVarPop}
\alias{.randFxVarPop}
\title{.randFxVarPop - random effects variance populator, see \code{\link{checkPolyICT}}}
\usage{
.randFxVarPop(phaseNames, groupNames, randFxVar, wn = c("p", "g", "n"))
}
\description{
.randFxVarPop - random effects variance populator, see \code{\link{checkPolyICT}}
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
