% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\docType{class}
\name{Err}
\alias{Err}
\title{\code{Err} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Err
}
\description{
\code{Err} class generator
}
\section{Fields}{

\describe{
\item{\code{model}}{The parameters for \code{FUN}.}

\item{\code{errVar}}{The error variance.}

\item{\code{FUN}}{The function for simulating error terms.}

\item{\code{fam}}{A random variable generator from the
\code{\link{gamlss.family}} for the error term distribution.
After data are simulated using \code{FUN} and its \code{model}, it can be
transformed to have a distribution implied by \code{fam}.}

\item{\code{famParms}}{Parameters to be passed to \code{fam}, see \code{\link{gamlss.family}}.}
}}

\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
