% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r, R/polyICT2.R
\name{expectedVar}
\alias{expectedVar}
\title{expectedVar}
\usage{
expectedVar(randFxCovMat, designMat, variances, randFxMean, nObservations, n)

expectedVar(randFxCovMat, designMat, variances, randFxMean, nObservations, n)
}
\description{
expectedVar

expectedVar
}
\author{
Stephen Tueller \email{stueller@rti.org}

Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
