% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTdesign.R
\docType{class}
\name{ICTdesign}
\alias{ICTdesign}
\title{\code{ICTdesign} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ICTdesign
}
\description{
\code{ICTdesign} class generator
}
\examples{
# create a basic ABA design
basicICT <- ICTdesign$new(n=10, phases=makePhase(), propErrVar=.5)
basicICT
basicICT$n <- 20
basicICT$n <- 'A'
basicICT$propErrVar <- 2
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
