% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{ICTpower}
\alias{ICTpower}
\title{ICTpower - get simulated power for an ICT design using parametric or
non-parametric bootstrap.}
\usage{
ICTpower(outFile = NULL, design = NULL, B = 100, dataFile = NULL,
  sampleSize = NULL, alpha = 0.05, seed = 123,
  cores = parallel::detectCores() - 1, savePowerReport = TRUE, ...)
}
\arguments{
\item{outFile}{The file name for saving power analysis results.
It may be a full path. The simulated data are to be saved by including the
file extension in a list. For example,
\code{file=c('n10_medSlopeEffect', 'csv')}
or \code{file=c('condition2', 'Rdata')}. Only `csv` and `RData` are supported.}

\item{design}{An \code{\link{designICT}} object such as
\code{\link{polyICT}}. This must be provided for a parametric bootstrap
esstimate of power.}

\item{B}{The number of simulated dataset (or parametric bootstrap replications).}

\item{dataFile}{Character. A file name for exntant data. Use this option to
do a non-parametric bootstrap (e.g., for finite sample power).
If \code{dataFile} is provide, \code{design}
will be ignored. The \code{dataFile} must be a *.csv file with the variables
'id' and 'Time', optionally 'phase' and/or 'group', and dependent variables
labeled 'y'. Alternatively, \code{dataFile} may be a *.Rdata file named 'Data'
with the same columns as described for the *.csv file.}

\item{sampleSize}{Numeric. The sample size to be drawn \code{B} times from
the \code{dataFile} data.}

\item{alpha}{Numeric. Default is .05. The Type I error rate for computing
emprical power (i.e., the proportion of p-values across \code{B} data sets
that are less than or equal to \code{alpha}).}

\item{seed}{A random seed for replicating the power analysis.}

\item{cores}{The number of cores used in parralelized simulation (for a
parametric bootstrap) and for fitting models to the data for both bootstrap
types (see \code{\link{PersonAlytic}}). The default
is to use one fewer cores than are detected on the computer. Do not exceed
the maximum available cores or unexpected results may occur and R may crash.}

\item{savePowerReport}{Should the power report be saved using the
\code{outFile} name? If \code{TRUE}, a `txt` file is saved in the working
directory. This option is set to \code{FALSE} in the function
\code{\link{ICTpowerSim}} which instead saves the reports from several
conditions in a `csv` file.}

\item{...}{Further arguments to be passed to \code{\link{PersonAlytic}} for
analysis. All options in \code{PersonAlytic} can be passed except for
\code{output}, \code{data}, \code{ids}, \code{dvs}, \code{time}, and
\code{phase} as these are taken from the \code{design}.}
}
\description{
ICTpower - get simulated power for an ICT design using parametric or
non-parametric bootstrap.
}
\examples{

example(polyICT)
myPolyICT$inputMat

\dontrun{

testICTpower10 <- ICTpower(c('testICTpower10', 'csv'),
  myPolyICT, B=100,
  seed = 54)
testICTpower20 <- ICTpower(c('testICTpower20', 'csv'),
  myPolyICT$update(groups=c(group1=20, group2=20)), B=100,
  seed = 23)
testICTpower20t100 <- ICTpower(c('testICTpower20', 'csv'),
  myPolyICT$update(groups=c(group1=20, group2=20), phases=makePhase(c(20,60,20))),
  B=100, seed = 23)

 }
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
