% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simICTdata.R
\name{ICTpowerSim}
\alias{ICTpowerSim}
\title{ICTpowerSim}
\usage{
ICTpowerSim(designs, save = NULL, B = 1000, seed = 1, alpha = 0.05,
  randFxFamily = "qNO", randFxParms = list(mu = 0.5, sigma = 1),
  errorParms = list(ar = c(0.5), ma = c(0)), errorFUN = arima.sim,
  errorFamily = NULL, cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{designs}{A named list of designs.
in \code{\link{ICTpower}}. The names in the named list will be passed to
the parameter \code{file} in \code{\link{ICTpower}}.}

\item{save}{Should the data be saved? If NULL (the default) raw data will not be
saved. Other options include \code{save='csv'} and \code{save='RData'}. This will be
combined with the names of \code{designs} to pass to the \code{file} option in
\code{\link{ICTpower}}.

The remaining parameters are as explained in \code{\link{ICTpower}}.}
}
\description{
ICTpowerSim
}
\examples{

designs <- list()
designs[["defaults"]] <- polyICT$new()
designs[["defaultsN20"]] <- designs[["defaults"]]$clone(deep=TRUE)
designs[["defaultsN20"]]$n <- 20
ICTpowerSim(designs, B=3)

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
