% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpowerSim.R
\name{ICTpowerSim}
\alias{ICTpowerSim}
\title{ICTpowerSim}
\usage{
ICTpowerSim(designs, pReportName = "ICTpowerSimResults", B = 1000,
  seed = 1, save = NULL, alpha = 0.05,
  cores = parallel::detectCores() - 1, ...)
}
\arguments{
\item{designs}{A named list of designs.
in \code{\link{ICTpower}}. The names in the named list will be passed to
the parameter \code{file} in \code{\link{ICTpower}}.}

\item{pReportName}{Character. A name used to create output files. The default
is \code{"ICTpowerSimResults"}}

\item{B}{Numeric (integer). The number of replicated data sets per \code{design}
in \code{designs}.}

\item{save}{\code{NULL} or character. Should the data be saved? If NULL
(the default) raw data will not be saved. Other
options include \code{save='csv'} and \code{save='RData'}. This will be
combined with the names of \code{designs} to pass to the \code{file} option
in \code{\link{ICTpower}}.

The remaining parameters are as explained in \code{\link{ICTpower}}.}

\item{Seed}{Numeric (integer). The random seed for ensuring the simulation
can be replicated.}
}
\description{
ICTpowerSim
}
\examples{

designs <- list()
designs[["defaults"]] <- polyICT$new()
designs[["defaultsN20"]] <- designs[["defaults"]]$clone(deep=TRUE)
designs[["defaultsN20"]]$n <- 20
ICTpowerSim(designs, B=3)

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
