% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makePhase}
\alias{makePhase}
\title{makePhase - a function to create the phase variable}
\usage{
makePhase(nObsPerPhase = c(10, 20, 10), phaseNames = c("A", "B", "A"))
}
\arguments{
\item{nObsPerPhase}{Numeric vector. The length is the number of phases
and each entry is the number of observations per phase.}

\item{phaseNames}{Numeric or character vector. The names of the phases,
e.g., phase 0 and phase 1, or phases "A", "B", and "C".}
}
\description{
makePhase - a function to create the phase variable
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
