% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makePhase}
\alias{makePhase}
\title{makePhase - a function to create the phase variable}
\usage{
makePhase(nObsPerCondition = c(10, 20, 10), conditions = c("A", "B",
  "A"))
}
\description{
makePhase - a function to create the phase variable
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
