% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDist.R
\name{makePhase}
\alias{makePhase}
\title{makePhase - function to make phases}
\usage{
makePhase(
  nObsPerPhase = c(10, 20, 10),
  phaseNames = c("phase1", "phase2", "phase3")
)
}
\arguments{
\item{nObsPerPhase}{Numeric vector. The length is the number of phases
and each entry is the number of observations per phase.}

\item{phaseNames}{Numeric or character vector. The names of the phases,
e.g., phase 0 and phase 1, or phases "A", "B", and "C". Phase names must
be unique, i.e., even if your design is ABA, you should use 1, 2, 3, or
"A", "B", "C".}
}
\value{
For \code{makePhase}, a list of phase names replicated by the number
of observations per phase.
}
\description{
makePhase - function to make phases
}
\examples{

# makePhase example
makePhase()
}
