% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTpower.R
\name{mse}
\alias{mse}
\title{mse - mean squared error & mean absolute error}
\usage{
mse(y, yhat)
}
\arguments{
\item{y}{Scalar. Population value of \code{y}.}

\item{yhat}{Numeric vector. Sample estimates of \code{y}.}
}
\description{
mse - mean squared error & mean absolute error
}
