% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyICT.r
\docType{class}
\name{polyICT}
\alias{polyICT}
\title{\code{polyICT} class generator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
polyICT
}
\description{
\code{polyICT} class generator
}
\section{Fields}{

\describe{
\item{\code{effectSizes}}{A length 2 list with the names `randFx` and `fixdFx`. Each
are themselves a list with the coefficients for a polynomial growth model.
The length of `randFx` and `fixdFx` corresponds to the order of the model,
where an order of 2 would be a linear model and an order of 3 would be a
quadratic model. An example of a quadratic model would be
list(randFx = list(intercept = 0, slope = .5, quad = .2),
     fixdFx = list(phase = 1, phaseSlope = .2, phaseQuad = .1))}

\item{\code{corMat}}{a symmetric correlation matrix with a dimension equal to the
order of the model. For example, a quadratic model with correpspond to a 3x3
matrix.}

\item{\code{randFxVar}}{a vector of the same length as the order of the polynomial
model containing the variances of the random effects. For example, in a
quadratic model, `randFxVar` would be length three, the first element would
be the intercept variance, the second element would be the slope variance,
and the third element would be the variance of the quadratic term.}
}}

\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{datasets}
